Normalize class when making HTML test comparisons
