+++ packages/perspective-viewer-d3fc/test/results/results.json
@@ -1,5 +1,5 @@
 {
-    "__GIT_COMMIT__": "f898b5486e7f2c6bf61f8648c8e6686cdee42535",
+    "__GIT_COMMIT__": "be50bcdfb1c1160ccc560af452db54446386ffcf",
     "area_shows_a_grid_without_any_settings_applied": "67bab9ea6654cc7cf7c7b096824e610a",
     "area_displays_visible_columns_": "9c195f0f7bf184a4cb9e727f1b37b301",
     "area_pivot_by_a_row": "fbae982b73c9cbf91be40a767cbbd3df",
@@ -121,28 +121,28 @@
     "sunburst_pivot_by_two_rows": "740a97d15d12ed2818e4e7404efa1290",
     "sunburst_pivot_by_a_row_and_a_column": "a9cc4064d026667a625ef9ca2fb927ff",
     "sunburst_pivot_by_two_rows_and_two_columns": "fe25573f98f12bfa788991c219978d35",
-    "treemap_pivot_by_a_row": "e499b4b4c91d9fdaafaf20d133b4943c",
-    "treemap_pivot_by_two_rows": "5e32d58d2070756dfa461b4eb1da9b7d",
-    "treemap_pivot_by_a_row_and_a_column": "e0d9e2504e068c9d6bb4519e0e6f2efe",
-    "treemap_pivot_by_two_rows_and_two_columns": "b61ebd6eb13fc292a95d5897af6edeb5",
-    "treemap_shows_a_grid_without_any_settings_applied": "10d1208b485425756fcc932229386b02",
-    "treemap_displays_visible_columns_": "10d1208b485425756fcc932229386b02",
-    "treemap_pivot_by_a_column": "10d1208b485425756fcc932229386b02",
-    "treemap_sort_by_a_hidden_column": "10d1208b485425756fcc932229386b02",
-    "treemap_sort_by_a_numeric_column": "10d1208b485425756fcc932229386b02",
-    "treemap_sort_by_an_alpha_column": "10d1208b485425756fcc932229386b02",
-    "treemap_filters_filters_by_a_numeric_column": "10d1208b485425756fcc932229386b02",
-    "treemap_filters_filters_by_an_alpha_column": "10d1208b485425756fcc932229386b02",
-    "treemap_filters_filters_with__in__comparator": "10d1208b485425756fcc932229386b02",
-    "sunburst_shows_a_grid_without_any_settings_applied": "10d1208b485425756fcc932229386b02",
-    "sunburst_displays_visible_columns_": "10d1208b485425756fcc932229386b02",
-    "sunburst_pivot_by_a_column": "10d1208b485425756fcc932229386b02",
-    "sunburst_sort_by_a_hidden_column": "10d1208b485425756fcc932229386b02",
-    "sunburst_sort_by_a_numeric_column": "10d1208b485425756fcc932229386b02",
-    "sunburst_sort_by_an_alpha_column": "10d1208b485425756fcc932229386b02",
-    "sunburst_filters_filters_by_a_numeric_column": "10d1208b485425756fcc932229386b02",
-    "sunburst_filters_filters_by_an_alpha_column": "10d1208b485425756fcc932229386b02",
-    "sunburst_filters_filters_with__in__comparator": "10d1208b485425756fcc932229386b02",
+    "treemap_pivot_by_a_row": "f39b569dba81f64090d60d767a5337de",
+    "treemap_pivot_by_two_rows": "646a5eb3ba2790f768f91b9681b59017",
+    "treemap_pivot_by_a_row_and_a_column": "95fb39573f9d966d7c4d3808d7fbae47",
+    "treemap_pivot_by_two_rows_and_two_columns": "9d848629342c6b541a3d949451232f05",
+    "treemap_shows_a_grid_without_any_settings_applied": "80705a191c3675ed34d3058874f09a00",
+    "treemap_displays_visible_columns_": "80705a191c3675ed34d3058874f09a00",
+    "treemap_pivot_by_a_column": "80705a191c3675ed34d3058874f09a00",
+    "treemap_sort_by_a_hidden_column": "80705a191c3675ed34d3058874f09a00",
+    "treemap_sort_by_a_numeric_column": "80705a191c3675ed34d3058874f09a00",
+    "treemap_sort_by_an_alpha_column": "80705a191c3675ed34d3058874f09a00",
+    "treemap_filters_filters_by_a_numeric_column": "80705a191c3675ed34d3058874f09a00",
+    "treemap_filters_filters_by_an_alpha_column": "80705a191c3675ed34d3058874f09a00",
+    "treemap_filters_filters_with__in__comparator": "80705a191c3675ed34d3058874f09a00",
+    "sunburst_shows_a_grid_without_any_settings_applied": "80705a191c3675ed34d3058874f09a00",
+    "sunburst_displays_visible_columns_": "80705a191c3675ed34d3058874f09a00",
+    "sunburst_pivot_by_a_column": "80705a191c3675ed34d3058874f09a00",
+    "sunburst_sort_by_a_hidden_column": "80705a191c3675ed34d3058874f09a00",
+    "sunburst_sort_by_a_numeric_column": "80705a191c3675ed34d3058874f09a00",
+    "sunburst_sort_by_an_alpha_column": "80705a191c3675ed34d3058874f09a00",
+    "sunburst_filters_filters_by_a_numeric_column": "80705a191c3675ed34d3058874f09a00",
+    "sunburst_filters_filters_by_an_alpha_column": "80705a191c3675ed34d3058874f09a00",
+    "sunburst_filters_filters_with__in__comparator": "80705a191c3675ed34d3058874f09a00",
     "bar_rendering_bugs_correctly_render_when_a_bar_chart_has_non_equidistant_times_on_a_datetime_axis": "96a10ce528c635ca24fa509e899bbb09",
     "events_perspective-config-update_event_is_fired_when_series_axis_is_changed": "c0cf24c214ad730393fdd7d58c843a8d",
     "events_perspective-config-update_event_is_fired_when_legend_position_is_changed": "9753be00c0d084dfbe592b46f6af6b08"
