+++ packages/perspective-viewer-datagrid/test/results/results.json
@@ -11,21 +11,21 @@
     "superstore_filters_by_a_datetime_column_": "4f3ebe78d2092c8c6f74389778d764d5",
     "superstore_highlights_invalid_filter_": "d74bde78c8afbd835e451fd5afffc9fe",
     "superstore_sorts_by_an_alpha_column_": "e404bb87320c31ec31a54c8652ae439c",
-    "superstore_displays_visible_columns_": "30dfb2e0551c4f598166959f5d7b3673",
+    "superstore_displays_visible_columns_": "09dd14f0e60da0e8749b7dd514e929f0",
     "superstore_resets_viewable_area_when_the_logical_size_expands_": "e9dd1d275f46f6a0857e0168835d0b38",
     "superstore_resets_viewable_area_when_the_physical_size_expands_": "e469597235cc032619bc095a76f9ad6f",
-    "__GIT_COMMIT__": "0ee79b07e87340cddb5626dfc732ca85932f6d7c",
-    "superstore_shows_a_grid_without_any_settings_applied": "8ec13290bdbbe3aa17405fc37fe771b7",
-    "superstore_pivot_by_a_row": "34f5a466415985ee75aeebcf3dc4f914",
-    "superstore_pivot_by_two_rows": "02dd963e15fa30e2a49a37fe3b606638",
-    "superstore_pivot_by_a_column": "13d1dbd9b8fa97a9053b111a93444d87",
-    "superstore_pivot_by_a_row_and_a_column": "f51e9465b6873f1b3e92aacb43ec89b8",
-    "superstore_pivot_by_two_rows_and_two_columns": "b5a05fc5031782f5b4d5d3f13dc2c966",
-    "superstore_sort_by_a_hidden_column": "80cfa97284910fa193e34a18286be57f",
-    "superstore_sort_by_a_numeric_column": "fa82c95d2f24795a11dfdaf7699511fb",
-    "superstore_sort_by_an_alpha_column": "e9f3d537de1c6283391623749974a2e8",
-    "superstore_filters_filters_by_a_numeric_column": "4e4f33833fb531bbdff6fe61694befd8",
-    "superstore_filters_filters_by_an_alpha_column": "2ee0b0c9c37540db277cfe6d390ebf2b",
-    "superstore_filters_filters_with__in__comparator": "414732a1eedbc4a3fcf070451e2aa7ed",
-    "superstore_perspective-config-update_event_is_fired_when_column_style_is_changed": "7885342bae75f6796eb231db2b741eeb"
+    "__GIT_COMMIT__": "be50bcdfb1c1160ccc560af452db54446386ffcf",
+    "superstore_shows_a_grid_without_any_settings_applied": "3ba1355b260e8ff927657844e7fdec4a",
+    "superstore_pivot_by_a_row": "709a4a6b40153d9bf013f90398ee568d",
+    "superstore_pivot_by_two_rows": "23db46ccf9d8327ae992fe3b5f468b3b",
+    "superstore_pivot_by_a_column": "12fe383947c0f5a91766d61a808ed6da",
+    "superstore_pivot_by_a_row_and_a_column": "93c241d67d1546e4284e78c1d34bb685",
+    "superstore_pivot_by_two_rows_and_two_columns": "5ecb4fcd02d839ba6b79c912eeabd1af",
+    "superstore_sort_by_a_hidden_column": "490c7fa58ef41fc0a43894f4ebd7dbe8",
+    "superstore_sort_by_a_numeric_column": "c8adc8030c989840a3febc2852cf6dda",
+    "superstore_sort_by_an_alpha_column": "d04e50eb057f3d682370b1803ff47a23",
+    "superstore_filters_filters_by_a_numeric_column": "6c6ba00b6edf9ee11f82848b7786fd91",
+    "superstore_filters_filters_by_an_alpha_column": "4b9a40c2ffad050ff4cf059e22cb2b68",
+    "superstore_filters_filters_with__in__comparator": "5882e2d2694519ba9880d0b2d5b19c44",
+    "superstore_perspective-config-update_event_is_fired_when_column_style_is_changed": "b9e0fabb3606d4ccc495efa481019842"
 }
