+++ packages/perspective-workspace/test/results/results.json
@@ -1,20 +1,20 @@
 {
-    "__GIT_COMMIT__": "c855e38b34e3ec514c2e8b61d1807ee89231e348",
+    "__GIT_COMMIT__": "be50bcdfb1c1160ccc560af452db54446386ffcf",
     "index_restore_workspace_with_detail_only": "d24f601369fbf86c853d4dd2894506e3",
-    "index_Light_DOM_restore_workspace_with_detail_only": "a86ed09f48f11f42776af77944fda2cf",
-    "index_Shadow_DOM_restore_workspace_with_detail_only": "087c2dffe30a52fa4ea1b127ce83c625",
-    "index_Light_DOM_restore_workspace_with_master_and_detail": "3fbd38e883ea2310419508607b81a69d",
-    "index_Shadow_DOM_restore_workspace_with_master_and_detail": "34a8933df88660871ef3ca3e1e2d6a11",
+    "index_Light_DOM_restore_workspace_with_detail_only": "9ae42ed9f3b0510bf2153b78f07ee062",
+    "index_Shadow_DOM_restore_workspace_with_detail_only": "aa44cdf6689ff93f68d4a4cffe292e09",
+    "index_Light_DOM_restore_workspace_with_master_and_detail": "b50b6f1b20190b6b33cf9ad9d222a6ee",
+    "index_Shadow_DOM_restore_workspace_with_master_and_detail": "221dcac5ed4a0e3b3ab1c1f72a6e8775",
     "index_Light_DOM_restore_workspace_with_viewers_with_generated_slotids": "0fad47e40121ecccf41c21bdfe2b7d56",
     "index_Shadow_DOM_restore_workspace_with_viewers_with_generated_slotids": "606bea2473f0e8b91756d5b69cd982be",
-    "index_Light_DOM_removeChild_Remove_One": "39ee7d3dd2cbff6f4ff9d960633b6fb2",
-    "index_Shadow_DOM_removeChild_Remove_One": "8f1a715acfc9a0415b9400a070dedfa4",
-    "index_Light_DOM_Create_One": "49010bc78a8a2375e57a2aa1627471ab",
-    "index_Light_DOM_Create_Multiple": "c8b96c9f843a022aa4f8b77a0a28312c",
-    "index_Light_DOM_Create_multiple_with_names": "a2dcd3235172a4921858273a52b5855a",
-    "index_Shadow_DOM_Create_One": "9731261b7a34c19fcb11855daf723c73",
-    "index_Shadow_DOM_Create_Multiple": "2776b343d54425badd479d57d300233e",
-    "index_Shadow_DOM_Create_multiple_with_names": "e8bd434eb1637078da674adf08e1d8dc",
+    "index_Light_DOM_removeChild_Remove_One": "2a1f7dd46b9aea1b1b98a79b59bb26e8",
+    "index_Shadow_DOM_removeChild_Remove_One": "423a6d5abe9dd037e15ce90b74c3a625",
+    "index_Light_DOM_Create_One": "4a73c047efa383c5b180d52e7ebfed0d",
+    "index_Light_DOM_Create_Multiple": "a646eaeee339aeadc061f5c88f76987c",
+    "index_Light_DOM_Create_multiple_with_names": "3aa508b53fe6266e394c6817ab5f6c02",
+    "index_Shadow_DOM_Create_One": "b0789b790a26ec4cd1fe095bd088e67c",
+    "index_Shadow_DOM_Create_Multiple": "4d88faf666989a462633244ece20f2c8",
+    "index_Shadow_DOM_Create_multiple_with_names": "391266f09bb098827941e474f3733680",
     "workspace-all_Light_DOM_Restore_workspace_\"Bucket_by_year": "c0bc13512d1449e3cde3f1a4c5963a70",
-    "workspace-all_Shadow_DOM_Restore_workspace_\"Bucket_by_year": "087c2dffe30a52fa4ea1b127ce83c625"
+    "workspace-all_Shadow_DOM_Restore_workspace_\"Bucket_by_year": "aa44cdf6689ff93f68d4a4cffe292e09"
 }
